Magnito/requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
python-decouple>=3.8
